+++ Capa_Presentacion/Modulos Funciones/CargaBitacora.vb
@@ -19,7 +19,7 @@ Module CargaBitacora
         EntidadBitacora.BaseDeDatos = BaseDeDatos
         NegocioBitacora.InsertaBitacora(EntidadBitacora)
     End Sub
-    Public Sub GeneraRegistroBitacora(ByVal Modulo As String, ByVal Opcion As String, Optional ByVal IdAdicional As Integer = 0, Optional ByVal ReferenciaAdicional As String = "")
+    Public Sub GeneraRegistroBitacora(ByVal Modulo As String, ByVal Opcion As String, Optional ByVal IdAdicional As Long = 0, Optional ByVal ReferenciaAdicional As String = "")
         Dim operacion As String = String.Empty
         Dim observaciones As String = String.Empty
         Try
